refactor(folder): enhance folder strcuture
