+++ src/utils/__init__.py
@@ -1,3 +1,1 @@
-from .logging_utils import logger
-from .paths import BasePath
 from .data_utils import _preprocess_text
